Storing datalists as globally accesible arrayss, so that csv files for these data lists only ever need be read once, or unless there needs to be an update to the data (eg, a new player, or a new team, like Tasmania).
Each time Index.html (and indirectly Layout.html) are rendered again in a user's sesion, the data lists will be populated from these global arrays, hence avoiding the need to access the csv files.

In a future exercise, a tool will be created such that the arrays are updated when the is a changeto the data in the csv files.